Support Toast notifications for async actions like `Delete` or `Restart`

+++ internal/pkg/views/notifications.go
@@ -25,6 +25,7 @@ type NotificationWidget struct {
 	x, y                          int
 	w                             int
 	pendingDeletes                []*expanders.TreeNode
+	toastNotifications            map[string]*eventing.StatusEvent
 	deleteMutex                   sync.Mutex // ensure delete occurs only once
 	deleteInProgress              bool
 	gui                           *gocui.Gui
@@ -35,7 +36,7 @@ type NotificationWidget struct {
 // delete once confirmed
 func (w *NotificationWidget) AddPendingDelete(item *expanders.TreeNode) {
 	if w.deleteInProgress {
-		eventing.SendStatusEvent(eventing.StatusEvent{
+		eventing.SendStatusEvent(&eventing.StatusEvent{
 			Failure: true,
 			Message: "Delete already in progress. Please wait for completion.",
 			Timeout: time.Second * 5,
@@ -44,7 +45,7 @@ func (w *NotificationWidget) AddPendingDelete(item *expanders.TreeNode) {
 	}
 
 	if item.DeleteURL == "" {
-		eventing.SendStatusEvent(eventing.StatusEvent{
+		eventing.SendStatusEvent(&eventing.StatusEvent{
 			Failure: true,
 			Message: "Item `" + item.Name + "` doesn't support delete",
 			Timeout: time.Second * 5,
@@ -56,7 +57,7 @@ func (w *NotificationWidget) AddPendingDelete(item *expanders.TreeNode) {
 	// current terminal size
 	_, yMax := w.gui.Size()
 	if len(w.pendingDeletes) > (yMax - 12) {
-		eventing.SendStatusEvent(eventing.StatusEvent{
+		eventing.SendStatusEvent(&eventing.StatusEvent{
 			Failure: true,
 			Message: "Can't add `" + item.Name + "` run out of space to draw the `Pending delete` list!",
 			Timeout: time.Second * 5,
@@ -69,7 +70,7 @@ func (w *NotificationWidget) AddPendingDelete(item *expanders.TreeNode) {
 
 	for _, i := range w.pendingDeletes {
 		if i.DeleteURL == item.DeleteURL {
-			eventing.SendStatusEvent(eventing.StatusEvent{
+			eventing.SendStatusEvent(&eventing.StatusEvent{
 				Failure: true,
 				Message: "Item already `" + item.Name + "` in pending delete list",
 				Timeout: time.Second * 5,
@@ -84,7 +85,7 @@ func (w *NotificationWidget) AddPendingDelete(item *expanders.TreeNode) {
 // ConfirmDelete delete all queued/pending deletes
 func (w *NotificationWidget) ConfirmDelete() {
 	if w.deleteInProgress {
-		eventing.SendStatusEvent(eventing.StatusEvent{
+		eventing.SendStatusEvent(&eventing.StatusEvent{
 			Failure: true,
 			Message: "Delete already in progress. Please wait for completion.",
 			Timeout: time.Second * 5,
@@ -110,7 +111,7 @@ func (w *NotificationWidget) ConfirmDelete() {
 			w.deleteInProgress = false
 		}()
 
-		event, _ := eventing.SendStatusEvent(eventing.StatusEvent{
+		event, _ := eventing.SendStatusEvent(&eventing.StatusEvent{
 			InProgress: true,
 			Message:    "Starting to delete items",
 			Timeout:    time.Second * 15,
@@ -153,6 +154,7 @@ func (w *NotificationWidget) ConfirmDelete() {
 
 		event.Message = "Delete request sent"
 		event.InProgress = false
+		event.SetTimeout(time.Second * 2)
 		event.Update()
 
 		w.pendingDeletes = []*expanders.TreeNode{}
@@ -164,7 +166,7 @@ func (w *NotificationWidget) ClearPendingDeletes() {
 	w.deleteMutex.Lock()
 	w.gui.Update(func(g *gocui.Gui) error {
 
-		_, done := eventing.SendStatusEvent(eventing.StatusEvent{
+		_, done := eventing.SendStatusEvent(&eventing.StatusEvent{
 			InProgress: true,
 			Message:    "Clearing pending deletes",
 			Timeout:    time.Second * 2,
@@ -181,26 +183,65 @@ func (w *NotificationWidget) ClearPendingDeletes() {
 // NewNotificationWidget create new instance and start go routine for spinner
 func NewNotificationWidget(x, y, w int, g *gocui.Gui, client *armclient.Client) *NotificationWidget {
 	widget := &NotificationWidget{
-		name:           "notificationWidget",
-		x:              x,
-		y:              y,
-		w:              w,
-		gui:            g,
-		pendingDeletes: []*expanders.TreeNode{},
-		client:         client,
+		name:               "notificationWidget",
+		x:                  x,
+		y:                  y,
+		w:                  w,
+		gui:                g,
+		pendingDeletes:     []*expanders.TreeNode{},
+		toastNotifications: map[string]*eventing.StatusEvent{},
+		client:             client,
 	}
+
+	newEvents := eventing.SubscribeToStatusEvents()
+	// Start loop for showing loading in statusbar
+	go func() {
+		// recover from panic, if one occurrs, and leave terminal usable
+		defer errorhandling.RecoveryWithCleanup()
+
+		for {
+			// Wait for a second to see if we have any new messages
+			timeout := time.After(time.Second)
+			select {
+			case eventObjRaw := <-newEvents:
+				eventObj := eventObjRaw.(*eventing.StatusEvent)
+				if eventObj.IsToast {
+					widget.toastNotifications[eventObj.ID()] = eventObj
+				}
+			case <-timeout:
+				// Update the UI
+			}
+			for ID, toast := range widget.toastNotifications {
+				if toast.HasExpired() {
+					delete(widget.toastNotifications, ID)
+				}
+			}
+
+			g.Update(func(gui *gocui.Gui) error {
+				return nil
+			})
+		}
+	}()
+
 	return widget
 }
 
 // Layout draws the widget in the gocui view
 func (w *NotificationWidget) Layout(g *gocui.Gui) error {
 	// Don't draw anything if no pending deletes
-	if len(w.pendingDeletes) < 1 {
+	if len(w.pendingDeletes) < 1 && len(w.toastNotifications) < 1 {
 		g.DeleteView(w.name)
 		return nil
 	}
 
-	height := len(w.pendingDeletes)*1 + 7
+	height := len(w.pendingDeletes) + len(w.toastNotifications)
+	if len(w.pendingDeletes) > 0 {
+		// Add padding for extra lines
+		height = height + 7
+	}
+	if len(w.toastNotifications) > 0 {
+		height = height + 3
+	}
 
 	v, err := g.SetView(w.name, w.x, w.y, w.x+w.w, height)
 	if err != nil && err != gocui.ErrUnknownView {
@@ -216,15 +257,26 @@ func (w *NotificationWidget) Layout(g *gocui.Gui) error {
 
 func (w *NotificationWidget) layoutInternal(v io.Writer) error {
 	pending := w.pendingDeletes
+	toasts := w.toastNotifications
 
-	fmt.Fprintln(v, style.Title("Pending Deletes:"))
-	for _, i := range pending {
-		fmt.Fprintln(v, " - "+i.Name)
+	if len(toasts) > 0 {
+		fmt.Fprintln(v, style.Title("Notifications:"))
+		for _, i := range toasts {
+			fmt.Fprintln(v, i.Icon()+" "+i.Message)
+		}
+		fmt.Fprintln(v, "")
 	}
-	fmt.Fprintln(v, "")
-	fmt.Fprintln(v, "Do you want to delete these items?")
-	fmt.Fprintln(v, style.Warning("Press "+strings.ToUpper(w.ConfirmDeleteKeyBinding)+" to DELETE"))
-	fmt.Fprintln(v, style.Highlight("Press "+strings.ToUpper(w.ClearPendingDeletesKeyBinding)+" to CANCEL"))
+
+	if len(pending) > 0 {
+		fmt.Fprintln(v, style.Title("Pending Deletes:"))
+		for _, i := range pending {
+			fmt.Fprintln(v, " - "+i.Name)
+		}
+		fmt.Fprintln(v, "")
+		fmt.Fprintln(v, "Do you want to delete these items?")
+		fmt.Fprintln(v, style.Warning("Press "+strings.ToUpper(w.ConfirmDeleteKeyBinding)+" to DELETE"))
+		fmt.Fprintln(v, style.Highlight("Press "+strings.ToUpper(w.ClearPendingDeletesKeyBinding)+" to CANCEL"))
+	}
 
 	return nil
 }
